Fix player level truncating to u16 when reading the wire, matches Canary fix
Matching client-side half of a server-side fix (separate Shaw-134/canary
fork): the server was sending level as a network uint16_t, silently
wrapping above 65535 (this project's endgame civilization starts at
350,000+ - a real level 174999 character displayed as 43927 in-game).
The server now sends a full uint32_t when a new ProtocolFeature::
PlayerDataLevelU32 flag is negotiated (always on for this project's
build) - this client needs the matching read width or every field after
a widened level field in the same packet would desync.

Adds Otc::GameLevelU32 (next free feature id after the real max,
GameProficiency=135), enabled unconditionally in
game_features/features.lua since it isn't tied to any real CipSoft
client version. Gates every level-shaped uint16_t read found via a full
getU16() pass over protocolgameparse.cpp: the main player-stats packet
(parsePlayerStats, previously ungated), the say/level display, both
Cyclopedia character-info packets (including a discarded-but-still-
stream-position-relevant read), the level-up client event, and the
highscores entry list.

Also widens the downstream storage types that would otherwise silently
re-truncate a correctly-read 32-bit value: LocalPlayer::m_level/
setLevel/getLevel, CyclopediaCharacterGeneralStats::level,
TaskBoardWeeklyHeaderData::currentPlayerLevel, Game::processTalk's
level param, and the highscores tuple's level slot (plus its
game.h/.cpp signatures) - mirrors this codebase's existing precedent
of widening m_experience/setExperience to uint64_t for GameDoubleExperience.

Left combat-skill levels (Fist, Sword, Forge stats, etc.) untouched -
a structurally different field, not implicated by this bug.

+++ src/client/localplayer.h
@@ -44,7 +44,7 @@ public:
     void setTotalCapacity(uint32_t totalCapacity);
     void setBaseCapacity(uint32_t baseCapacity);
     void setExperience(uint64_t experience);
-    void setLevel(uint16_t level, uint16_t levelPercent);
+    void setLevel(uint32_t level, uint16_t levelPercent);
     void setMana(uint32_t mana, uint32_t maxMana);
     void setManaShield(uint32_t manaShield, uint32_t maxManaShield);
     void setMagicLevel(uint16_t magicLevel, uint16_t magicLevelPercent);
@@ -86,7 +86,7 @@ public:
     uint16_t getLevelPercent();
     uint8_t getHarmony() { return m_harmony; }
 
-    uint16_t getLevel() { return m_level; }
+    uint32_t getLevel() { return m_level; }
     uint16_t getSkillLevel(const Otc::Skill skill) { return m_skills[skill].level; }
     uint16_t getSkillBaseLevel(const Otc::Skill skill) { return m_skills[skill].baseLevel; }
     uint16_t getSkillLevelPercent(const Otc::Skill skill) { return m_skills[skill].levelPercent; }
@@ -211,7 +211,7 @@ private:
     uint32_t m_health{ 0 };
     uint32_t m_maxHealth{ 0 };
     uint64_t m_experience{ 0 };
-    uint16_t m_level{ 0 };
+    uint32_t m_level{ 0 };
     uint16_t m_levelPercent{ 0 };
     uint32_t m_mana{ 0 };
     uint32_t m_maxMana{ 0 };
