+++ src/client/staticdata.h
@@ -133,7 +133,7 @@ struct TaskBoardPreferredSlotData
 struct TaskBoardWeeklyHeaderData
 {
     uint8_t difficulty{ 0 };
-    uint16_t currentPlayerLevel{ 0 };
+    uint32_t currentPlayerLevel{ 0 };
     uint8_t remainingDays{ 0 };
     uint8_t totalTaskSlots{ 0 };
     uint32_t maxExperience{ 0 };
@@ -482,7 +482,7 @@ struct StoreData
 struct CyclopediaCharacterGeneralStats
 {
     uint64_t experience;
-    uint16_t level;
+    uint32_t level;
     uint16_t levelPercent;
     uint16_t baseExpGain;
     uint16_t lowLevelExpBonus;
